requirements.txt
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
diskcache==5.6.3
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
mongoengine==0.29.1
numpy==2.3.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.13.2
requests==2.32.4
setuptools==78.1.1
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
wheel==0.45.1
